Added the genquery command
The command makes query files from a selection menu. Also restructured
the paths to a description of the downloaded and processed columns.
Added a type parameter to the interfaces to distinguish the different
data types. Each interface type now has to pivot around a single common
column. This allows the dependencies to be simpler - it is the
interface's problem to get the required data to have the pivot column of
interest. I hope I won't regret this simple design later on...

I will, won't I?